refactor(gateway): improve modules separation
